data_server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
yfinance>=0.2.40
pydantic>=2.0.0
